control-plane/orchestrator/requirements.txt
psycopg2-binary>=2.9
grpcio>=1.50
grpcio-tools>=1.50
protobuf>=4.21
